requirements.txt
```bash
mlx
mlx-lm
numpy
Pillow
tqdm